fix: fix up tests broken by the reactor/proactor move
