+++ tests/Paramore.Brighter.Redis.Tests/MessagingGateway/Proactor/When_a_message_consumer_throws_a_socket_exception_when_connecting_to_the_server_async.cs
@@ -1,42 +1,37 @@
 ﻿using System;
+using System.Threading.Tasks;
 using FluentAssertions;
 using Paramore.Brighter.MessagingGateway.Redis;
 using Paramore.Brighter.Redis.Tests.TestDoubles;
 using ServiceStack.Redis;
 using Xunit;
 
-namespace Paramore.Brighter.Redis.Tests.MessagingGateway;
+namespace Paramore.Brighter.Redis.Tests.MessagingGateway.Proactor;
 
 [Collection("Redis Shared Pool")]   //shared connection pool so run sequentially
 [Trait("Category", "Redis")]
-public class RedisMessageConsumerRedisNotAvailableTests : IDisposable
+public class RedisMessageConsumerRedisNotAvailableTestsAsync 
 {
     private readonly ChannelName _queueName = new ChannelName("test");
     private readonly RoutingKey _topic = new RoutingKey("test");
     private readonly RedisMessageConsumer _messageConsumer;
     private Exception? _exception;
 
-    public RedisMessageConsumerRedisNotAvailableTests()
+    public RedisMessageConsumerRedisNotAvailableTestsAsync()
     {
         var configuration = RedisFixture.RedisMessagingGatewayConfiguration();
 
         _messageConsumer = new RedisMessageConsumerSocketErrorOnGetClient(configuration, _queueName, _topic);
-
     }
 
     [Fact]
-    public void When_a_message_consumer_throws_a_socket_exception_when_connecting_to_the_server()
+    public async Task When_a_message_consumer_throws_a_socket_exception_when_connecting_to_the_server_async()
     {
-        _exception = Catch.Exception(() => _messageConsumer.Receive(TimeSpan.FromMilliseconds(1000))); 
-            
+        _exception = await Catch.ExceptionAsync(() => _messageConsumer.ReceiveAsync(TimeSpan.FromMilliseconds(1000)));
+
         _exception.Should().BeOfType<ChannelFailureException>();
         _exception?.InnerException.Should().BeOfType<RedisException>();
-  
     }
-        
-    public void Dispose()
-    {
-        _messageConsumer.Purge();
-        _messageConsumer.Dispose();
-    }
+
+//Do not dispose of fake client
 }
