+++ tests/Paramore.Brighter.Redis.Tests/TestDoubles/RedisMessageConsumerSocketErrorOnGetClient.cs
@@ -1,4 +1,6 @@
 ﻿using System.Net.Sockets;
+using System.Threading;
+using System.Threading.Tasks;
 using Paramore.Brighter.MessagingGateway.Redis;
 using ServiceStack.Redis;
 
@@ -15,7 +17,11 @@ public class RedisMessageConsumerSocketErrorOnGetClient(
 
     protected override IRedisClient GetClient()
     {
-        throw new RedisException(SocketException, new SocketException((int) SocketError.AccessDenied));
+        throw new ChannelFailureException("Simulated socked exception", new RedisException(SocketException, new SocketException((int) SocketError.AccessDenied)));
     }
 
+    protected override Task<IRedisClientAsync?> GetClientAsync(CancellationToken cancellationToken = default(CancellationToken))
+    {
+        throw new ChannelFailureException("Simulated socked exception", new RedisException(SocketException, new SocketException((int) SocketError.AccessDenied)));
+    }
 }
